requirements.txt
langchain 
langchain-chroma 
langchain-openai 
langchain-huggingface
beautifulsoup4
langchain-community
faiss-cpu
langchain-text-splitters
langchain-experimental
langchain-chroma
requests
python-dotenv
langchainhub
tavily-python
gradio
gradio_client
beautifulsoup4
pypdf
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fsspec==2025.5.1
gradio==5.36.2
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.10.1
mcpadapt==0.1.10
mdurl==0.1.2
nltk==3.9.1
numpy==2.3.1
openai==1.95.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.3
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smolagents==1.20.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.1
textblob==0.19.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yfinance==0.2.27
finnhub-python
plotly
mlflow 
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
# google-cloud-aiplatform
# google-cloud-bigquery
# google-cloud-core
# google-cloud-resource-manager
# google-cloud-storage
# google-crc32c
google-generativeai
# google-resumable-media
# googleapis-common-protos
# kfp
vertexai
protobuf
pyarrow
google-adk
litellm
streamlit
load_dotenv